+++ package.json
@@ -4,49 +4,24 @@
   "description": "",
   "keywords": [],
   "main": "src/index.tsx",
+  "license": "MIT",
   "dependencies": {
-    "@tiptap/core": "2.0.0-beta.174",
-    "@tiptap/extension-bubble-menu": "2.0.0-beta.55",
-    "@tiptap/extension-code-block-lowlight": "2.0.0-beta.68",
-    "@tiptap/extension-document": "2.0.0-beta.15",
-    "@tiptap/extension-link": "2.0.0-beta.36",
-    "@tiptap/extension-placeholder": "2.0.0-beta.48",
-    "@tiptap/extension-task-item": "2.0.0-beta.31",
-    "@tiptap/extension-task-list": "2.0.0-beta.26",
-    "@tiptap/extension-typography": "2.0.0-beta.20",
-    "@tiptap/html": "2.0.0-beta.173",
-    "@tiptap/react": "2.0.0-beta.114",
-    "@tiptap/starter-kit": "2.0.0-beta.183",
-    "@tiptap/suggestion": "2.0.0-beta.90",
-    "classnames": "2.3.1",
-    "dompurify": "2.3.3",
-    "emojibase": "5.2.0",
-    "emojibase-data": "6.2.0",
-    "emojibase-regex": "5.1.3",
-    "lodash-es": "4.17.21",
-    "marked": "3.0.8",
-    "overlayscrollbars": "1.13.1",
-    "overlayscrollbars-react": "0.2.3",
+    "@tiptap/core": "^2.0.0-beta.182",
+    "@tiptap/extension-bubble-menu": "^2.0.0-beta.61",
+    "@tiptap/extension-document": "^2.0.0-beta.17",
+    "@tiptap/html": "^2.0.0-beta.181",
+    "@tiptap/react": "^2.0.0-beta.114",
+    "@tiptap/starter-kit": "^2.0.0-beta.191",
     "react": "17.0.2",
-    "react-cool-inview": "2.0.7",
     "react-dom": "17.0.2",
     "react-icons": "4.3.1",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.3",
-    "rehype": "12.0.0",
-    "rehype-format": "4.0.0",
-    "turndown": "7.1.1",
-    "twemoji-parser": "13.1.0"
+    "react-scripts": "4.0.3"
   },
   "devDependencies": {
-    "@types/dompurify": "2.3.1",
-    "@types/lodash-es": "4.17.5",
-    "@types/marked": "3.0.1",
     "@types/react": "17.0.20",
     "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.3.1",
-    "@types/turndown": "5.0.1",
-    "@types/twemoji-parser": "13.1.1",
+    "@typescript-eslint/parser": "^5.33.1",
+    "sass": "^1.54.4",
     "typescript": "4.4.2"
   },
   "scripts": {
