+++ src/components/RichTextEditor.scss
@@ -1,3 +1,4 @@
+
 .ProseMirror {
     &:focus-visible {
         outline: none;
