+++ src/global.d.ts
@@ -1,0 +1,9 @@
+// declare module "@tiptap/core" {
+//   interface Commands<ReturnType> {
+//     columnBlock: {
+//       setColumns: (columns: number) => ReturnType;
+//       insertColumns: (columns: number) => ReturnType;
+//       unsetColumns: () => ReturnType;
+//     };
+//   }
+// }
