+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "react-dom";
 
-import { Tiptap } from "./components";
+import { RichTextEditor } from "./components";
 
 import { BASIC_CONTENT } from "./data";
 
@@ -9,12 +9,8 @@ import "./index.scss";
 
 render(
   <div id="Wrapper">
-    <Tiptap
+    <RichTextEditor
       content={BASIC_CONTENT}
-      withToolbar={true}
-      withTaskListExtension={true}
-      withLinkExtension={true}
-      withEmojisReplacer={true}
     />
   </div>,
   document.getElementById("root")
