+++ src/react-app-env.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="react-scripts" />
+
+import { Commands } from '@tiptap/core'
+
+declare module "@tiptap/core" {
+
+  interface Commands<ReturnType> {
+    columnBlock: {
+      setColumns: (columns: number) => ReturnType;
+      insertColumns: (columns: number) => ReturnType;
+      unsetColumns: () => ReturnType;
+    };
+  }
+}
